Expose register and rename signal API
