+++ django_durable/management/commands/durable_signal.py
@@ -2,7 +2,7 @@ import json
 
 from django.core.management.base import BaseCommand, CommandError
 
-from django_durable.engine import send_signal
+from django_durable.engine import signal_workflow
 from django_durable.models import WorkflowExecution
 
 
@@ -31,5 +31,5 @@ class Command(BaseCommand):
         except WorkflowExecution.DoesNotExist:
             raise CommandError(f'WorkflowExecution not found: {exec_id}')
 
-        send_signal(wf, name, payload)
+        signal_workflow(wf, name, payload)
         self.stdout.write(self.style.SUCCESS('OK'))
